#485: Move TimeZoneRulesCommon and RecurringZoneRules implementations to TimeZoneRules.kt
